Add Outlet and Layout
