+++ src/App.tsx
@@ -1,25 +1,8 @@
 import { MouseEvent, useState } from 'react'
 import Button from './components/Button/Button'
 import Input from './components/Input/Input'
-import { RouterProvider, createBrowserRouter } from 'react-router-dom'
-import { Menu } from './pages/Menu'
-import { Cart } from './pages/Cart'
-import { ErrorPage } from './pages/ErrorPage'
+import { Link } from 'react-router-dom'
 
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: <Menu />
-  },
-  {
-    path: '/cart',
-    element: <Cart />
-  },
-  {
-    path: "*",
-    element: <ErrorPage />
-  }
-])
 
 function App() {
   const [counter, setCounter] = useState<number>(0)
@@ -30,14 +13,11 @@ function App() {
 
   return (
     <>
-      <Button onClick={addCounter}>Кнопка</Button>
-      <Button appearence='big' onClick={addCounter}>Кнопка</Button>
-      <Input placeholder='Email' />
-      <div>
-        <a href="/">Menu</a>
-        <a href="/cart">Cart</a>
+      {/* <div>
+        <Link to="/">Menu</Link>
+        <Link to="/cart">Cart</Link>
       </div>
-      <RouterProvider router={router} />
+       */}
     </>
   )
 }
