+++ src/pages/Cart.tsx
@@ -1,3 +1,6 @@
+import { Link } from "react-router-dom";
+
 export function Cart() {
-    return <>Cart</>
+    return <>Cart
+    </>
 }
