+++ src/pages/Menu.tsx
@@ -1,3 +1,5 @@
 export function Menu() {
-    return <>Menu</>
+    return <>
+    Menu
+    </>
 }
